DifferentiableStateSpaceModels

Documentation for DifferentiableStateSpaceModels.

DifferentiableStateSpaceModels.@include_example_module

DifferentiableStateSpaceModels.@include_example_module — Macro

Return model from a stored module, creating as necessary . If the module is

already included, it just returns the Module

saved to disk but not already included, it does include

not saved to disk, it saves it, includes it

The specification of a model_generator has a sepcification like H, mod_vals = model_generator() where H is an symbolics expression the mod_vals can be splatted into the PerturbationModel constructor.

The name of the stored module is the name of model_generator function and the order.

Examples

``julia-repl julia> m = @include_example_module(Examples.rbc_observables_benchmark) # stores or loads.functioncache/rbcobservables.jl`